Allow reloading account.account from CSV files
Work around code in `account.account` that forbids writing to certain
fields even if they are unchanged by not trying to write to them in
that case.

This allows updating modules that contain accounts in CSV files, since
CSV files don't allow specifying noupdate.

This is temporary workaround until the PR below is accepted:
https://github.com/odoo/odoo/pull/16188